requirements.txt
numpy>=1.21.0
torch>=1.9.0
mne>=1.0.0
scipy>=1.7.0
scikit-learn>=0.24.2
wandb>=0.12.0
matplotlib>=3.4.0
pandas>=1.3.0
pytest>=6.2.5
tqdm>=4.62.0
pillow>=8.3.0 